Botão de confirmação

+++ src/app/usuarios/containers/usuarios/usuarios.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 
 import { Usuario } from '../../model/usuario';
@@ -11,10 +12,9 @@ import { UsuariosService } from '../../services/usuarios.service';
 @Component({
   selector: 'app-usuarios',
   templateUrl: './usuarios.component.html',
-  styleUrls: ['./usuarios.component.scss']
+  styleUrls: ['./usuarios.component.scss'],
 })
 export class UsuariosComponent implements OnInit {
-
   usuarios$: Observable<Usuario[]> | null = null;
 
   // UsuariosService: UsuariosService;
@@ -25,30 +25,26 @@ export class UsuariosComponent implements OnInit {
     public router: Router,
     private route: ActivatedRoute,
     private snackBar: MatSnackBar
-
   ) {
     this.refresh();
   }
 
   refresh() {
-    this.usuarios$ = this.UsuariosService.list()
-      .pipe(
-        catchError(error => {
-          this.onError('Erro no carregamento dos usuários.')
-          return of([])
-        })
-      );
+    this.usuarios$ = this.UsuariosService.list().pipe(
+      catchError((error) => {
+        this.onError('Erro no carregamento dos usuários.');
+        return of([]);
+      })
+    );
   }
 
   onError(errorMsg: string) {
     this.dialog.open(ErrorDialogComponent, {
-      data: errorMsg
+      data: errorMsg,
     });
   }
 
-  ngOnInit(): void {
-
-  }
+  ngOnInit(): void {}
 
   onAdd() {
     this.router.navigate(['new'], { relativeTo: this.route });
@@ -59,12 +55,24 @@ export class UsuariosComponent implements OnInit {
   }
 
   onRemove(usuario: Usuario) {
-    this.UsuariosService.remove(usuario._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Usuário excluído com sucesso!', 'X', { duration: 5000, verticalPosition: 'top', horizontalPosition: 'center' })
-      },
-      () => this.onError('Erro ao tentar excluir o usuário.')
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja excluir este usuário?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.UsuariosService.remove(usuario._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Usuário excluído com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          () => this.onError('Erro ao tentar excluir o usuário.')
+        );
+      }
+    });
   }
 }
